novas atividades

+++ python/beecrowd/1044.py
@@ -1,4 +1,4 @@
-a, b = map(float, input().split())
+a, b = map(int, input().split())
 
 if a % b == 0 or b % a == 0:
     print('Sao Multiplos')
